fix: renamed directories
